+++ usvg/src/text/mod.rs
@@ -125,6 +125,7 @@ fn text_to_paths(
         shaper::apply_writing_mode(writing_mode, &mut clusters);
         shaper::apply_letter_spacing(chunk, &mut clusters);
         shaper::apply_word_spacing(chunk, &mut clusters);
+        shaper::apply_length_adjust(chunk, &mut clusters);
         let mut curr_pos = shaper::resolve_clusters_positions(
             chunk,
             char_offset,